+++ netbox_healthcheck_plugin/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Arthur Hanson"""
 __email__ = 'ahanson@netboxlabs.com'
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
 
 from netbox.plugins import PluginConfig
@@ -12,7 +12,7 @@ class HealthCheckConfig(PluginConfig):
     name = 'netbox_healthcheck_plugin'
     verbose_name = 'NetBox HealthCheck Plugin'
     description = 'NetBox plugin for HealthCheck.'
-    version = 'version'
+    version = '0.3.0'
     base_url = 'netbox_healthcheck_plugin'
     django_apps = [
         'health_check',
@@ -20,6 +20,8 @@ class HealthCheckConfig(PluginConfig):
         'health_check.contrib.migrations',
         'health_check.contrib.redis',
     ]
-    min_version = "v4.0-beta1"
+    min_version = '4.5.0'
+    max_version = '4.99.99'
+
 
 config = HealthCheckConfig